Time: 0 ms (100%), Space: 7.5 MB (78.91%) - LeetHub

+++ 2050-count-good-numbers/2050-count-good-numbers.cpp
@@ -1,7 +1,7 @@
 #define mod 1000000007;
 class Solution {
 public:
-    long long power(int x, int n){
+    long long power(long long x, long long n){
         if(n==0){
             return 1;
         }
